pyproject.toml
[build-system]
requires = ["maturin>=1.0,<2.0"]
build-backend = "maturin"

# NOTE: This section is a stub; needed to prevent
# `uv run --only-group docs` from failing.
[project]
name = "zizmor"
dynamic = ["version"]
# Arbitrarily set to the oldest non-EOL Python.
requires-python = ">=3.9"

[tool.maturin]
bindings = "bin"
manifest-path = "crates/zizmor/Cargo.toml"

[dependency-groups]
docs = ["mkdocs ~= 1.6", "mkdocs-material[imaging] ~= 9.5"]
codegen = ["prance[osv]", "requests", "pyyaml"]
